일단 user+VBoard+Comment 합치는 과정 중

+++ src/main/java/com/vaccinelife/vaccinelifeapi/service/CommentService.java
@@ -7,6 +7,7 @@ import com.vaccinelife.vaccinelifeapi.model.VBoard;
 import com.vaccinelife.vaccinelifeapi.repository.CommentRepository;
 import com.vaccinelife.vaccinelifeapi.repository.UserRepository;
 import com.vaccinelife.vaccinelifeapi.repository.VBoardRepository;
+import lombok.AllArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -27,33 +28,33 @@ public class CommentService {
         VBoard vBoard = vBoardRepository.findById(requestDto.getVBoardId()).orElseThrow(
                 () -> new IllegalArgumentException("해당 게시물이 존재하지 않습니다.")
         );
-//        User user = userRepository.findByUserId(requestDto.getUserId()).orElseThrow(
-//                () -> new IllegalArgumentException("해당 아이디가 존재하지 안습니다.")
-//        );
+        User user = userRepository.findByUserId(requestDto.getUserId()).orElseThrow(
+                () -> new IllegalArgumentException("해당 아이디가 존재하지 안습니다.")
+        );
         Comment comment = Comment.builder()
-//                .user(user)
+                .user(user)
                 .vBoard(vBoard)
                 .comment(requestDto.getComment()).build();
         commentRepository.save(comment);
     }
-//
-//    @Transactional(readOnly = true)
-//    public List<CommentRequestDto> getComment() {
-//        List<Comment> comments = commentRepository.findAll();
-//        return CommentRequestDto.list(comments);
-//    }
+
+    @Transactional(readOnly = true)
+    public List<CommentRequestDto> getComment() {
+        List<Comment> comments = commentRepository.findAll();
+        return CommentRequestDto.list(comments);
+    }
 
     @Transactional
     public void deleteComment(Long id, CommentRequestDto requestDto) {
-//        User user = userRepository.findByUserId(requestDto.getUserId()).orElseThrow(
-//                () -> new IllegalArgumentException("해당 유저를 찾을 수 없습니다.")
-//        );
+        User user = userRepository.findByUserId(requestDto.getUserId()).orElseThrow(
+                () -> new IllegalArgumentException("해당 유저를 찾을 수 없습니다.")
+        );
         Comment comment = commentRepository.findById(id).orElseThrow(
                 () -> new IllegalArgumentException("해당 댓글을 찾을 수 없습니다.")
         );
-//        if(!comment.getUser().equals(user)){
-//            throw new IllegalArgumentException("자신이 쓴 댓글만 삭제 할 수 있습니다.");
-//        }
+        if(!comment.getUser().equals(user)){
+            throw new IllegalArgumentException("자신이 쓴 댓글만 삭제 할 수 있습니다.");
+        }
         commentRepository.delete(comment);
